requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv